requirements.txt
pyserial~=3.5
RPi.GPIO
Flask~=2.2.2